adiciona verificacao de remocao e validacao de campos do form

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -6,6 +6,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -50,17 +51,25 @@ export class CoursesComponent {
     }
 
     onRemove(course: Course) {
-      this.coursesService.remove(course._id).subscribe({
-        next: data => {
-          this.refresh();
-          this._snackBar.open("Curso removido com sucesso","X",{
-            duration:5000,
-            verticalPosition: 'top',
-            horizontalPosition: 'center'
-          });
-        },
-        error: error => this.onError('Erro ao tentar remover curso')
-      })
-    }
 
+      const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+        data: 'Tem certeza que gostaria de remover o curso?'
+      });
+
+      dialogRef.afterClosed().subscribe((result: boolean) => {
+        if (result) {
+          this.coursesService.remove(course._id).subscribe({
+            next: data => {
+              this.refresh();
+              this._snackBar.open("Curso removido com sucesso","X",{
+                duration:5000,
+                verticalPosition: 'top',
+                horizontalPosition: 'center'
+              });
+            },
+            error: error => this.onError('Erro ao tentar remover curso')
+          });
+        }
+      });
+    }
 }
